Fix/package structure (#331)
* Move __init__.py content to component.py

* Export factory and mix elements from __init__.py

* Move version variable and VideoTransformerFactory alias to __init__.py

* Move docstring to __init__.py

* Format
