+++ Makefile
@@ -1,5 +1,5 @@
 pkg/build:
-	python release_check.py streamlit_webrtc/__init__.py
+	python release_check.py streamlit_webrtc/component.py
 	cd streamlit_webrtc/frontend && npm run build
 	poetry build
